requirements.txt
dotmap==1.3.30
faster_whisper==1.1.0
numpy==2.2.0
opencv_python_headless==4.10.0.84
Pillow==11.0.0
